backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.4.2
httpx==0.25.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
redis==5.0.1
websockets==12.0
openai==1.3.5
tiktoken==0.5.1
firebase-admin==6.2.0
google-cloud-firestore==2.13.1 